[DocDB] Removing custom 'IfNotExists' code as DocumentClient now supports this
