+++ src/WebJobs.Extensions.DocumentDB/DocumentDBUtility.cs
@@ -2,7 +2,9 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
+using System.Threading.Tasks;
 using Microsoft.Azure.Documents;
+using Microsoft.Azure.Documents.Client;
 
 namespace Microsoft.Azure.WebJobs.Extensions.DocumentDB
 {
@@ -26,6 +28,48 @@ namespace Microsoft.Azure.WebJobs.Extensions.DocumentDB
             documentClientEx = ae.InnerException as DocumentClientException;
 
             return documentClientEx != null;
+        }
+
+        internal static async Task CreateDatabaseAndCollectionIfNotExistAsync(DocumentDBContext context)
+        {
+            await CreateDatabaseAndCollectionIfNotExistAsync(context.Service, context.ResolvedAttribute.DatabaseName, context.ResolvedAttribute.CollectionName,
+                context.ResolvedAttribute.PartitionKey, context.ResolvedAttribute.CollectionThroughput);
+        }
+
+        internal static async Task CreateDatabaseAndCollectionIfNotExistAsync(IDocumentDBService service, string databaseName, string collectionName, string partitionKey, int throughput)
+        {
+            await service.CreateDatabaseIfNotExistsAsync(new Database { Id = databaseName });
+
+            await CreateDocumentCollectionIfNotExistsAsync(service, databaseName, collectionName, partitionKey, throughput);
+        }
+
+        private static async Task<DocumentCollection> CreateDocumentCollectionIfNotExistsAsync(IDocumentDBService service, string databaseName, string collectionName,
+            string partitionKey, int throughput)
+        {
+            Uri databaseUri = UriFactory.CreateDatabaseUri(databaseName);
+
+            DocumentCollection documentCollection = new DocumentCollection
+            {
+                Id = collectionName
+            };
+
+            if (!string.IsNullOrEmpty(partitionKey))
+            {
+                documentCollection.PartitionKey.Paths.Add(partitionKey);
+            }
+
+            // If there is any throughput specified, pass it on. DocumentClient will throw with a 
+            // descriptive message if the value does not meet the collection requirements.
+            RequestOptions collectionOptions = null;
+            if (throughput != 0)
+            {
+                collectionOptions = new RequestOptions
+                {
+                    OfferThroughput = throughput
+                };
+            }
+
+            return await service.CreateDocumentCollectionIfNotExistsAsync(databaseUri, documentCollection, collectionOptions);
         }
     }
 }
