+++ test/WebJobs.Extensions.Tests/Extensions/DocumentDB/DocumentDBTestUtility.cs
@@ -13,22 +13,76 @@ using Microsoft.Azure.Documents.Client;
 using Microsoft.Azure.WebJobs.Extensions.DocumentDB;
 using Microsoft.Azure.WebJobs.Extensions.Tests.Extensions.DocumentDB.Models;
 using Microsoft.Azure.WebJobs.Host.Bindings;
+using Moq;
 using Newtonsoft.Json.Linq;
 
 namespace Microsoft.Azure.WebJobs.Extensions.Tests.Extensions.DocumentDB
 {
     internal static class DocumentDBTestUtility
     {
+        public const string DatabaseName = "ItemDB";
+        public const string CollectionName = "ItemCollection";
+
+        public static void SetupCollectionMock(Mock<IDocumentDBService> mockService, string partitionKeyPath = null, int throughput = 0)
+        {
+            Uri databaseUri = UriFactory.CreateDatabaseUri(DatabaseName);
+
+            var expectedPaths = new List<string>();
+            if (!string.IsNullOrEmpty(partitionKeyPath))
+            {
+                expectedPaths.Add(partitionKeyPath);
+            }
+
+            if (throughput == 0)
+            {
+                mockService
+                    .Setup(m => m.CreateDocumentCollectionIfNotExistsAsync(databaseUri,
+                        It.Is<DocumentCollection>(d => d.Id == CollectionName && Enumerable.SequenceEqual(d.PartitionKey.Paths, expectedPaths)),
+                        null))
+                    .ReturnsAsync(new DocumentCollection());
+            }
+            else
+            {
+                mockService
+                    .Setup(m => m.CreateDocumentCollectionIfNotExistsAsync(databaseUri,
+                        It.Is<DocumentCollection>(d => d.Id == CollectionName && Enumerable.SequenceEqual(d.PartitionKey.Paths, expectedPaths)),
+                        It.Is<RequestOptions>(r => r.OfferThroughput == throughput)))
+                    .ReturnsAsync(new DocumentCollection());
+            }
+        }
+
+        public static void SetupDatabaseMock(Mock<IDocumentDBService> mockService)
+        {
+            mockService
+                .Setup(m => m.CreateDatabaseIfNotExistsAsync(It.Is<Database>(d => d.Id == DatabaseName)))
+                .ReturnsAsync(new Database());
+        }
+
+        public static DocumentDBContext CreateContext(IDocumentDBService service, bool createIfNotExists = false,
+            string partitionKeyPath = null, int throughput = 0)
+        {
+            DocumentDBAttribute attribute = new DocumentDBAttribute(DocumentDBTestUtility.DatabaseName, DocumentDBTestUtility.CollectionName)
+            {
+                CreateIfNotExists = createIfNotExists,
+                PartitionKey = partitionKeyPath,
+                CollectionThroughput = throughput
+            };
+
+            return new DocumentDBContext
+            {
+                Service = service,
+                ResolvedAttribute = attribute
+            };
+        }
+
         public static IOrderedQueryable<T> AsOrderedQueryable<T, TKey>(this IEnumerable<T> enumerable, Expression<Func<T, TKey>> keySelector)
         {
             return enumerable.AsQueryable().OrderBy(keySelector);
         }
 
-        public static DocumentClientException CreateDocumentClientException(HttpStatusCode status, int retryAfter = 0)
+        public static DocumentClientException CreateDocumentClientException(HttpStatusCode status)
         {
             var headers = new NameValueCollection();
-            headers.Add("x-ms-retry-after-ms", retryAfter.ToString());
-
             var parameters = new object[] { null, null, headers, status, null };
             return Activator.CreateInstance(typeof(DocumentClientException), BindingFlags.NonPublic | BindingFlags.Instance, null, parameters, null) as DocumentClientException;
         }
